Structure repository according to the Arduino Library Specification
A popular library installation technique is to download the library via GitHub's Clone or download > Download ZIP and then use the Arduino IDE's Sketch > Include Library > Add .ZIP Library on the downloaded file. This requires the library to be in the root of the repository, not in a subfolder. If the library is not in the root of the repository this installation technique fails:

Specified folder/zip file does not contain a valid library

This move is also required if you wanted to add your library to the Arduino Library Manager index, which provides an even easier installation option.

Sketches placed inside the library (in the examples folder per the specification) will appear under the Arduino IDE's File > Examples > ardvga menu after installing the library. Sketches must be in a folder that matches the filename of the primary .ino file.

Reference:
https://github.com/arduino/Arduino/wiki/Arduino-IDE-1.5:-Library-specification